fixed incorrect casting

+++ src/include/phenology.h
@@ -12,7 +12,7 @@ double  calc_ncd(double);
 double  leaf_drop(double, double, double);
 void    calc_ini_grass_pheno_stuff(control *, met *, int, double *, double *,
                                 double *);
-void    calculate_growing_season_fluxes(fluxes *f, state *, double);
+void    calculate_growing_season_fluxes(fluxes *f, state *, int);
 void    calculate_days_left_in_growing_season(control *, state *, int, int, int);
